Parse safecopy with serialize fallback
I determined all of our database writes go through encode/decode
functions that already do compression. I have it now on decode try
safecopy first and failing that, go back to serialize for old
values. This is because there appears to be no way to write a safecopy
instance that does not encode a version tag first *unless* it is a
primitive which does not allow for forward migrations which defeats
the purpose of using safecopy in this case. We should be safe now to
do an upgrade without any sor of flushing needing to happen.