Dimension field spike
I'm not sure about this. We either need to basically treat dimensions
as a further bit of specificity over just metric names or we need to
store it on the actual counter/sampler. I think the problem with this
approach I've taken is you don't get your in-memory aggregation across
the same metric with different dimensions. However, putting it on the
counter/sampler doesn't make sense because its totally possible to
sampleI/counterI on differing dimensions.

Maybe we need a 2-level map and then we can combine the stats for
multiple dimension sets into one, but also emit them separately as
separate submission packets?