Spike on configuring workers to control quantiles
We'll see how this plays out, but the idea here is to allow each
worker to decide which quantiles get published. This is especially
critical for cloudwatch, which does not allow you to submit quantile
metrics directly. So that means we need to create sub-metrics for each
quantile and because of how heavy that is, it is really important to
allow the user to curtail the quantiles that get published.

I tried to keep things simple and just take a map from metric name to
quantile set, but I think this is not appropriate for other
backends. Other backends (like graphite IIRC), encode information into
their metric names rather than dimensions. This means that there's a
runtime component to the metric name. For those backends, if they
wanted to use this feature and not just use standardQuantiles, they
would need a function to match the shape of a metric key, rather than
just specify it fully with a map. I tried to provide some combinators
to make this simpler.