Make cloudwatch worker thread more resilient.
Currently, when you start a cloudwatch agg process it starts up a worker thread
and on each metric event it pushes to a bounded queue. If the worker hits a
logical exception like bad credentials when pushing metrics, it will quietly die
and the queue will fill until it hits the configured size limit and then will
just drop metrics on the floor. This change adds an error hook so you can log or
do what you need to do on error before the loop runs again. If your callback has
a synchronous error, it will be swallowed. It also adds an optional configurable
delay on error so you can avoid flooding the log if the error persists.